docker/requirements.txt
flask==3.0.3
Pillow==9.5.0
torch
torchvision
torchaudio
numpy
scipy
tqdm==4.67.1
safetensors==0.4.2
transformers==4.30.2
typing-extensions==4.12.2
sympy==1.13.3
networkx==3.3
filelock==3.13.1
jinja2==3.1.3
fsspec==2024.6.1
diffusers==0.34.0
accelerate==1.9.0
huggingface_hub==0.34.3
opencv-python-headless==4.10.0.84
soundfile
librosa
numba
resampy
pydub
imageio
imageio-ffmpeg
face_alignment
yacs
pyyaml
joblib
scikit-image
basicsr
facexlib
gradio
gfpgan
av
anyascii
aiohttp
packaging
mutagen
pysbd
umap-learn
pandas
matplotlib
jieba
pypinyin
hangul_romanize
gruut[de,es,fr]==2.2.3
jamo
nltk
g2pkk
bangla
bnnumerizer
bnunicodenormalizer
einops
unidecode
num2words
spacy[ja]
so-vits-svc-fork==4.2.30